# Conflicts:
#	gradle.properties
#	src/main/java/slimeknights/tconstruct/fluids/fluids/PotionFluidAttributes.java
#	src/main/java/slimeknights/tconstruct/library/fluid/FluidTooltipHandler.java
#	src/main/java/slimeknights/tconstruct/smeltery/block/entity/component/TankBlockEntity.java
#	src/main/java/slimeknights/tconstruct/smeltery/block/entity/controller/MelterBlockEntity.java
#	src/main/java/slimeknights/tconstruct/smeltery/block/entity/module/MeltingModuleInventory.java

+++ src/main/java/slimeknights/tconstruct/library/client/data/AbstractFluidTooltipProvider.java
@@ -109,7 +109,7 @@ public abstract class AbstractFluidTooltipProvider extends GenericDataProvider {
     }
 
     /** Adds a unit local to the current mod */
-    public FluidUnitListBuilder addTinkerUnit(String key, int amount) {
+    public FluidUnitListBuilder addTinkerUnit(String key, long amount) {
       return addUnitRaw(TConstruct.makeTranslationKey("gui", "fluid." + key), amount);
     }
